requirements.txt
flask
transformers
pillow